requirements.txt
# requirements.txt — Ejercicio 1: CNN + Gradio + Twilio
# Instala con: pip install -r requirements.txt

tensorflow==2.16.1
gradio==4.44.0
twilio==9.3.0
numpy==1.26.4
Pillow==10.4.0
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
